added user checks in comment methods

+++ src/main/java/com/example/flightadvisor/service/impl/CityServiceImpl.java
@@ -2,12 +2,15 @@ package com.example.flightadvisor.service.impl;
 
 import com.example.flightadvisor.exceptions.CityNotFoundException;
 import com.example.flightadvisor.exceptions.CommentNotFoundException;
+import com.example.flightadvisor.exceptions.InvalidArgumentsException;
 import com.example.flightadvisor.model.City;
 import com.example.flightadvisor.model.Comment;
 import com.example.flightadvisor.repository.CityRepository;
 import com.example.flightadvisor.repository.CommentRepository;
 import com.example.flightadvisor.service.CityService;
 import com.example.flightadvisor.service.CommentService;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -33,11 +36,6 @@ public class CityServiceImpl implements CityService {
     }
 
     @Override
-    public Optional<City> findById(Long cityId) {
-        return this.cityRepository.findById(cityId);
-    }
-
-    @Override
     public Optional<City> findByName(String name) {
         return this.cityRepository.findByNameLike(name);
     }
@@ -54,7 +52,7 @@ public class CityServiceImpl implements CityService {
     public Optional<City> addCommentForCity(Long cityId, Comment comment) {
         City city = this.cityRepository.findById(cityId).orElseThrow(() -> new CityNotFoundException(cityId));
 
-        Comment commentToAdd = new Comment(comment.getDescription());
+        Comment commentToAdd = new Comment(comment.getDescription(), comment.getCreator());
         this.commentRepository.save(commentToAdd);
 
         city.getComments().add(commentToAdd);
@@ -62,21 +60,29 @@ public class CityServiceImpl implements CityService {
     }
 
     @Override
-    public Optional<City> editCommentOfCity(Long cityId, Long commentId, Comment comment) {
+    public Optional<City> editCommentOfCity(Long cityId, Long commentId, Comment comment, String modifier) {
         City city = this.cityRepository.findById(cityId).orElseThrow(() -> new CityNotFoundException(cityId));
+        Comment commentToEdit = this.commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException(commentId));
 
-        this.commentService.editComment(commentId, comment.getDescription());
-        return Optional.of(city);
+        if(modifier.equals(commentToEdit.getCreator())){
+            this.commentService.editComment(commentId, comment.getDescription());
+            return Optional.of(city);
+        } else
+            throw new InvalidArgumentsException();
     }
 
     @Override
-    public Optional<City> deleteCommentOfCity(Long cityId, Long commentId) {
+    public Optional<City> deleteCommentOfCity(Long cityId, Long commentId, String username) {
         City city = this.cityRepository.findById(cityId).orElseThrow(() -> new CityNotFoundException(cityId));
         Comment commentToDelete = this.commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException(commentId));
-        city.getComments().remove(commentToDelete);
 
-        this.commentRepository.delete(commentToDelete);
-        return Optional.of(city);
+        if(username.equals(commentToDelete.getCreator())){
+            city.getComments().remove(commentToDelete);
+
+            this.commentRepository.delete(commentToDelete);
+            return Optional.of(city);
+        }else
+            throw new InvalidArgumentsException();
     }
